Fixed new util
